irefighting robotic system
This firefighting robotic system is powered by Arduino
Uno development board it consists of the HC-SR04 ultra-
sonic sensor mounted on a servo motor for obstacles
detection and free path navigation, it is also equipped
with the fire flame sensor for detecting and approaching
fire it also makes use of water tank and spray mechanism
for extinguishing the fire. Water spraying nozzle is
mounted on servo motor to cover maximum area. Water
is pumped from the main water tank to the water nozzle
with the help of water pump.